requirements.txt
Django==3.1.5
django-environ==0.4.5
django-widget-tweaks==1.4.8
mysqlclient==2.0.3
django-allauth==0.44.0
#psycopg2-binary==2.8.6
#sentry-sdk==0.19.5

# For Production only
gunicorn==20.0.4
#django-cities-light==3.8.1
